Fixed bugs in tick generators. These were caused by the fact that MAX_VALUE-MIN_VALUE doesn't fit in its defining type, so I must have changed the difference type to a larger type.

+++ src/main/java/org/esa/beam/dataViewer3D/data/axis/tickgenerator/_byte/TickCountTickGenerator.java
@@ -48,7 +48,7 @@ public class TickCountTickGenerator extends
         tickLabels[0] = min.toString();
         tickLabels[numTicks + 1] = max.toString();
 
-        byte difference = (byte) (max - min); // ok, max is always bigger
+        int difference = max - min;
 
         // if we have more ticks than integers between min and max, then we have to create less ticks
         if (difference < numTicks)
